+++ service-worker.js
@@ -1,4 +1,4 @@
-const CACHE_NAME = "forge-mvp-v126";
+const CACHE_NAME = "forge-mvp-v127";
 const CORE_ASSETS = [
   "./",
   "./request-help/",